Move utils/machine -> machine
