+++ container/docker/docker.go
@@ -24,7 +24,7 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/google/cadvisor/info/v1"
-	"github.com/google/cadvisor/utils/machine"
+	"github.com/google/cadvisor/machine"
 )
 
 func Status() (v1.DockerStatus, error) {
