+++ machine/testdata/cpuinfo
