+++ manager/manager.go
@@ -36,8 +36,8 @@ import (
 	"github.com/google/cadvisor/fs"
 	info "github.com/google/cadvisor/info/v1"
 	"github.com/google/cadvisor/info/v2"
+	"github.com/google/cadvisor/machine"
 	"github.com/google/cadvisor/utils/cpuload"
-	"github.com/google/cadvisor/utils/machine"
 	"github.com/google/cadvisor/utils/oomparser"
 	"github.com/google/cadvisor/utils/sysfs"
 	"github.com/google/cadvisor/version"
